requirements.txt
python-dotenv
requests
gspread